Combined remove/add method from search

+++ MovieMatchMvc/Controllers/MovieController.cs
@@ -43,14 +43,6 @@ namespace MovieMatchMvc.Controllers
 			return View("Watchlist", model);
 		}
 
-		[HttpPost]
-		[Route("AddMovieToList")]
-		public async Task<IActionResult> AddMovieToList(int movieId)
-		{
-			string userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-			await _movieService.AddMovieToWatchlistById(movieId, userId);
-			return Json(new { success = true });
-		}
 
 		[HttpGet("MatchWatchLists")]
 		public IActionResult MatchWatchLists()
@@ -87,15 +79,40 @@ namespace MovieMatchMvc.Controllers
 			await _movieService.RemoveFromWatchListAsync(movieId, userId);
 			return RedirectToAction(nameof(Watchlist));
 		}
+
 		[HttpPost]
-		[Route("RemoveFromWatchListSearch")]
-		public async Task<IActionResult> RemoveFromWatchListSearch(int movieId)
+		[Route("ManageWatchList")]
+		public async Task<IActionResult> ManageWatchList(int movieId, bool remove)
 		{
-			Console.WriteLine($"Received Movie ID: {movieId} to be removed");
 			string userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-			await _movieService.RemoveFromWatchListAsync(movieId, userId);
+			if (remove)
+			{
+				await _movieService.RemoveFromWatchListAsync(movieId, userId);
+			}
+			else
+			{
+				await _movieService.AddMovieToWatchlistById(movieId, userId);
+			}
 			return Json(new { success = true });
 		}
 
+		//[HttpPost]
+		//[Route("AddMovieToList")]
+		//public async Task<IActionResult> AddMovieToList(int movieId)
+		//{
+		//	string userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+		//	await _movieService.AddMovieToWatchlistById(movieId, userId);
+		//	return Json(new { success = true });
+		//}
+		//------------
+		//[HttpPost]
+		//[Route("RemoveFromWatchListSearch")]
+		//public async Task<IActionResult> RemoveFromWatchListSearch(int movieId)
+		//{
+		//	Console.WriteLine($"Received Movie ID: {movieId} to be removed");
+		//	string userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+		//	await _movieService.RemoveFromWatchListAsync(movieId, userId);
+		//	return Json(new { success = true });
+		//}
 	}
 }
